+++ src/exec/img_utils.c
@@ -21,7 +21,7 @@ static void	set_pix_img(t_imgdata *img, int x, int y, int color)
 		*(int *)target = color;
 }
 
-static float	get_perc_face(int face, t_coord *point)
+static double	get_perc_face(int face, t_coord *point)
 {
 	if (face == 1)
 		return (point->x - (int)(point->x));
